requirements.txt
# Chat Tool - OpenAI API聊天系统依赖包
# 基于实际部署环境生成 - 2025-07-12

# Web Framework & Server
fastapi==0.116.1              # FastAPI Web框架
uvicorn==0.35.0               # ASGI服务器
starlette==0.47.1             # FastAPI依赖的ASGI框架

# OpenAI Integration
openai==1.95.1                # OpenAI官方Python客户端

# Data Validation & Models
pydantic==2.11.7              # 数据验证和序列化
pydantic-core==2.33.2         # Pydantic核心库

# Template Engine
jinja2==3.1.6                 # 模板引擎
markupsafe==3.0.2             # Jinja2安全标记依赖

# Environment & Configuration
python-dotenv==1.1.1          # 环境变量加载

# File Operations
aiofiles==24.1.0              # 异步文件操作
python-multipart==0.0.20      # 表单数据处理

# HTTP & Network
httpx==0.28.1                 # HTTP客户端 (OpenAI依赖)
httpcore==1.0.9               # HTTP核心库
h11==0.16.0                   # HTTP/1.1协议实现
certifi==2025.7.9             # SSL证书验证

# Type Annotations
typing-extensions==4.14.1     # 类型注解扩展
annotated-types==0.7.0        # 注解类型支持

# Utilities
click==8.2.1                  # 命令行工具 (uvicorn依赖)
anyio==4.9.0                  # 异步IO库
sniffio==1.3.1                # 异步库检测
idna==3.10                    # 国际化域名支持
tqdm==4.67.1                  # 进度条 (OpenAI依赖)
distro==1.9.0                 # Linux发行版检测 (OpenAI依赖)
jiter==0.10.0                 # JSON迭代器 (Pydantic依赖)
